Restructure: extract Journey + TimingError + TimedLeg into journey.rs
